Fix trim example
The trim example should feature the "start" point of the new audio clip and the clip's length, rather than the clip's end time. According to the sox manual page, the following example creates a 30 second clip, starting from the beginning of a file:

sox song.wav ringtone.wav trim 0 30

I have tested this and confirmed the man page is correct.